server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
paho-mqtt

# Image Processing & ML
opencv-python
ultralytics
numpy